// src/types/tmdbSearch.ts
export type tmdbSearch = {
  id: string;
  name: string;
  genre_ids: number[];
  backdrop_path: string;
  adult: boolean;
  origin_country: string[];
  original_language: string;
  original_name: string;
  overview: string;
  popularity: number;
  poster_path: string;
  first_air_date: string;
  vote_average: number;
  vote_count: number;
};